requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
ffmpeg-python==0.2.0
filelock==3.12.0
future==0.18.3
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.1rc1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper==20230314
pycparser==2.21
regex==2023.6.3
requests==2.31.0
scipy==1.10.1
sounddevice==0.4.6
SpeechRecognition==3.10.0
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
